Use make and stylus
Note: right now the stylus code is exactly the same as the css was.  We can
keep it that way (and just use the features we want, like transparent mixins
and variables), or we can use more stylus features.  Either way I think we
should put what we're planning in the README (or, better yet, a CONTRIBUTING
file) so developers know what the plan is.

Note: I've added a Node .gitignore (thanks
https://github.com/github/gitignore) to avoid committing things like
node_modules/ which show up now because of the cli tools we're using now from
node